basic layout of how our robot will function
This code can drive the robot, and run the intake and shooting motors. Needs to be polished, but the robot can intake balls from the human player station and shoot. Robot can also intake from ground, but it still needs an actual intake. Mecanum drivetrain is still broken, it is very important that we fix that. Slowing down front motors to match gear ratio did not fix it.